Get GPS position with a simple loop

+++ app/src/main/java/com/example/dops_location_app/model/LocationResponse.kt
@@ -1,3 +1,6 @@
 package com.example.dops_location_app.model
 
-data class LocationResponse(var latitude : Float, val longitude : Float, val elapsedTime: Long)
+data class LocationResponse(
+    var latitude : Double,
+    val longitude : Double,
+    val elapsedTime: Double)
